Logging and improved error reporting
* Moved case class/objects to separate files. This improves clarity and structure.
* Fixed the use of `null` in comparison expressions. The solution was to wrap the `null` in
an `Option` and then checking if the opbject `isDefined`.
* Implemented logging with typesafe/scalalogging library
* Failed transactions now properly log error messages returned from the Neo4j server.
* Methods that execute transactions now return `false` on failure.
* Upped version to 0.8.2
